refactor: setStatusToPackageByMarketplace and setStatusToReadyToShip
