+++ src/Service/BaseOrderManager.php
@@ -420,113 +420,6 @@ class BaseOrderManager extends BaseManager
         return $this->getOrders($parameters);
     }
 
-    /**
-     * @param mixed[] $orderItemIds
-     *
-     * @return OrderItem[]
-     */
-    public function setStatusToPackedByMarketplace(
-        array $orderItemIds,
-        string $deliveryType,
-        string $shippingProvider = null,
-        string $trackingNumber = null
-    ): array {
-        $action = 'SetStatusToPackedByMarketplace';
-
-        $parameters = $this->makeParametersForAction($action);
-        $parameters->set([
-            'OrderItemIds' => Json::encode($orderItemIds),
-            'DeliveryType' => $deliveryType,
-        ]);
-
-        if (!empty($shippingProvider)) {
-            $parameters->set(['ShippingProvider' => $shippingProvider]);
-        }
-
-        if (!empty($trackingNumber)) {
-            $parameters->set(['TrackingNumber' => $trackingNumber]);
-        }
-
-        $requestId = $this->generateRequestId();
-
-        $builtResponse = $this->executeAction(
-            $action,
-            $parameters,
-            $requestId,
-            'POST'
-        );
-
-        $orderItems = OrderItemsFactory::makeFromStatus($builtResponse->getBody());
-
-        $orderItemsResponse = array_values($orderItems->all());
-
-        $this->logger->info(
-            sprintf(
-                '%d::%s::APIResponse::SellerCenterSdk: the items status was changed',
-                $requestId,
-                $action
-            )
-        );
-
-        return $orderItemsResponse;
-    }
-
-    /**
-     * @param mixed[] $orderItemIds
-     *
-     * @return OrderItem[]
-     */
-    public function setStatusToReadyToShip(
-        array $orderItemIds,
-        string $deliveryType,
-        string $shippingProvider = null,
-        string $trackingNumber = null,
-        ?string $packageId = null
-    ): array {
-        $action = 'SetStatusToReadyToShip';
-
-        $parameters = $this->makeParametersForAction($action);
-        $parameters->set([
-            'OrderItemIds' => Json::encode($orderItemIds),
-            'DeliveryType' => $deliveryType,
-        ]);
-
-        if (!empty($shippingProvider)) {
-            $parameters->set(['ShippingProvider' => $shippingProvider]);
-        }
-
-        if (!empty($trackingNumber)) {
-            $parameters->set(['TrackingNumber' => $trackingNumber]);
-        }
-
-        if (!empty($packageId)) {
-            $parameters->set(['PackageId' => $packageId]);
-        }
-
-        $requestId = $this->generateRequestId();
-
-        $builtResponse = $this->executeAction(
-            $action,
-            $parameters,
-            $requestId,
-            'POST'
-        );
-
-        $orderItems = OrderItemsFactory::makeFromStatus($builtResponse->getBody());
-
-        $orderItemsResponse = array_values($orderItems->all());
-
-        $this->logger->info(
-            sprintf(
-                '%d::%s::APIResponse::SellerCenterSdk: the items status was changed',
-                $requestId,
-                $action
-            )
-        );
-
-        return $orderItemsResponse;
-    }
-
     public function getTrackingCode(
         string $packageId,
         string $shippingProvider
