+++ src/Service/GlobalOrderManager.php
@@ -5,6 +5,8 @@ declare(strict_types=1);
 namespace Linio\SellerCenter\Service;
 
 use Linio\SellerCenter\Response\SuccessResponse;
+use Linio\Component\Util\Json;
+use Linio\SellerCenter\Factory\Xml\Order\OrderItemsFactory;
 class GlobalOrderManager extends BaseOrderManager
 {
     public function setInvoiceNumber(
@@ -42,4 +44,92 @@ class GlobalOrderManager extends BaseOrderManager
 
         return $response;
     }
+
+       /**
+     * @param mixed[] $orderItemIds
+     *
+     * @return OrderItem[]
+     */
+    public function setStatusToReadyToShip(
+        array $orderItemIds,
+        string $deliveryType,
+        ?string $packageId = null
+    ): array {
+        $action = 'SetStatusToReadyToShip';
+
+        $parameters = $this->makeParametersForAction($action);
+        $parameters->set([
+            'OrderItemIds' => Json::encode($orderItemIds),
+            'DeliveryType' => $deliveryType,
+        ]);
+
+        if (!empty($packageId)) {
+            $parameters->set(['PackageId' => $packageId]);
+        }
+
+        $requestId = $this->generateRequestId();
+
+        $builtResponse = $this->executeAction(
+            $action,
+            $parameters,
+            $requestId,
+            'POST'
+        );
+
+        $orderItems = OrderItemsFactory::makeFromStatus($builtResponse->getBody());
+
+        $orderItemsResponse = array_values($orderItems->all());
+
+        $this->logger->info(
+            sprintf(
+                '%d::%s::APIResponse::SellerCenterSdk: the items status was changed',
+                $requestId,
+                $action
+            )
+        );
+
+        return $orderItemsResponse;
+    }
+
+    /**
+     * @param mixed[] $orderItemIds
+     *
+     * @return OrderItem[]
+     */
+    public function setStatusToPackedByMarketplace(
+        array $orderItemIds,
+        string $deliveryType
+    ): array {
+        $action = 'SetStatusToPackedByMarketplace';
+
+        $parameters = $this->makeParametersForAction($action);
+        $parameters->set([
+            'OrderItemIds' => Json::encode($orderItemIds),
+            'DeliveryType' => $deliveryType,
+        ]);
+
+        $requestId = $this->generateRequestId();
+
+        $builtResponse = $this->executeAction(
+            $action,
+            $parameters,
+            $requestId,
+            'POST'
+        );
+
+        $orderItems = OrderItemsFactory::makeFromStatus($builtResponse->getBody());
+
+        $orderItemsResponse = array_values($orderItems->all());
+
+        $this->logger->info(
+            sprintf(
+                '%d::%s::APIResponse::SellerCenterSdk: the items status was changed',
+                $requestId,
+                $action
+            )
+        );
+
+        return $orderItemsResponse;
+    }
+
 }
